Refactor file locations again
