+++ decorators/nameable.rb
